requirements/develop.txt
-r base.txt
django-debug-toolbar==3.5.0
django_extensions==3.2.0
django-override-storage==0.3.2
django-test-without-migrations==0.6
django-webtest==1.9.10
factory-boy==3.2.1
mock==4.0.3
pyfakefs==4.5.6
pytest==7.1.2
pytest-clarity==1.0.1
pytest-cov==3.0.0
pytest-diff==0.1.14
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-forked==1.4.0
pytest-freezegun==0.4.2
pytest-icdiff==0.5
pytest-mock==3.8.1
pytest-raises==0.11
pytest-randomly==3.12.0
pytest-sugar==0.9.4
pytest-xdist==2.5.0
selenium==4.3.0
WebTest==3.0.0
